Add DocumentMetadataMapper
